replace CRA with Parcel in material-ui-starter
