+++ packages/starter/material-ui-starter/.prettierignore
@@ -1,1 +1,5 @@
-build
+.parcel-cache
+dist
+docs
+node_modules
+
