+++ packages/starter/material-ui-starter/.prettierrc
@@ -1,0 +1,7 @@
+{
+    "printWidth": 120,
+    "trailingComma": "es5",
+    "tabWidth": 4,
+    "semi": true,
+    "singleQuote": true
+}
