--- packages/starter/material-ui-starter/craco.config.js
@@ -1,1 +1,0 @@
-module.exports = {};
