+++ packages/starter/material-ui-starter/package.json
@@ -1,68 +1,65 @@
 {
     "name": "@solana/wallet-adapter-material-ui-starter",
-    "version": "0.12.1",
+    "version": "0.13.0",
     "author": "Solana Maintainers <maintainers@solana.foundation>",
     "repository": "https://github.com/solana-labs/wallet-adapter",
     "license": "Apache-2.0",
-    "files": [
-        "public",
-        "src",
-        "craco.config.js",
-        "LICENSE",
-        "README.md",
-        "tsconfig.json"
-    ],
     "publishConfig": {
         "access": "public"
     },
+    "files": [
+        "src",
+        ".editorconfig",
+        ".eslintignore",
+        ".eslintrc.json",
+        ".gitignore",
+        ".prettierignore",
+        ".prettierrc",
+        "LICENSE",
+        "package.json",
+        "tsconfig.json"
+    ],
     "scripts": {
-        "clean": "shx rm -rf build/*",
-        "start": "craco start",
-        "build": "yarn clean && craco build",
-        "test": "craco test",
-        "analyze": "source-map-explorer 'build/static/js/*.js'"
+        "clean": "shx rm -rf dist .parcel-cache",
+        "build": "yarn clean && parcel build src/index.html",
+        "start": "yarn clean && parcel src/index.html",
+        "fmt": "prettier --write '{*,**/*}.{js,ts,jsx,tsx,json}'",
+        "lint": "eslint . && prettier --check '{*,**/*}.{js,ts,jsx,tsx,json}'",
+        "lint:fix": "eslint --fix . && yarn fmt",
+        "nuke": "shx rm -rf node_modules yarn.lock"
     },
     "dependencies": {
         "@emotion/react": "^11.0.0",
         "@emotion/styled": "^11.0.0",
         "@mui/icons-material": "^5.0.0",
         "@mui/material": "^5.0.0",
-        "@solana/wallet-adapter-base": "^0.9.1",
+        "@solana/wallet-adapter-base": "^0.9.2",
+        "@solana/wallet-adapter-react": "^0.15.2",
         "@solana/wallet-adapter-material-ui": "^0.16.0",
-        "@solana/wallet-adapter-react": "^0.15.1",
-        "@solana/wallet-adapter-wallets": "^0.14.1",
-        "@solana/web3.js": "^1.20.0",
+        "@solana/wallet-adapter-wallets": "^0.14.2",
+        "@solana/web3.js": "^1.31.0",
         "notistack": "^2.0.0",
         "react": "^17.0.0",
         "react-dom": "^17.0.0"
     },
     "devDependencies": {
-        "@craco/craco": "^6.0.0",
-        "@types/node": "^12.0.0",
-        "@types/react": "^17.0.0",
-        "@types/react-dom": "^17.0.0",
-        "eslint-config-react-app": "^6.0.0",
-        "react-scripts": "^4.0.0",
-        "source-map-explorer": "^2.5.2",
-        "typescript": "~4.4.4",
-        "yarn": "^1.22.0"
-    },
-    "eslintConfig": {
-        "extends": [
-            "react-app",
-            "react-app/jest"
-        ]
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
+        "@types/eslint": "^8.2.2",
+        "@types/eslint-plugin-prettier": "^3.1.0",
+        "@types/node": "^16.11.14",
+        "@types/prettier": "^2.4.3",
+        "@types/react": "^17.0.38",
+        "@types/react-dom": "^17.0.11",
+        "@typescript-eslint/eslint-plugin": "^5.9.1",
+        "@typescript-eslint/parser": "^5.9.1",
+        "eslint": "^8.6.0",
+        "eslint-config-prettier": "^8.3.0",
+        "eslint-plugin-prettier": "^4.0.0",
+        "eslint-plugin-react": "^7.28.0",
+        "eslint-plugin-react-hooks": "^4.3.0",
+        "parcel": "^2.2.0",
+        "prettier": "^2.5.1",
+        "shx": "^0.3.4",
+        "tslib": "^2.3.1",
+        "typescript": "^4.5.4"
     }
 }
