+++ packages/starter/material-ui-starter/src/App.tsx
@@ -1,54 +1,74 @@
-import { createTheme, StyledEngineProvider, ThemeProvider } from '@mui/material';
-import { deepPurple } from '@mui/material/colors';
-import { SnackbarProvider } from 'notistack';
-import React, { FC } from 'react';
-import { Wallet } from './Wallet';
+import { WalletAdapterNetwork, WalletError } from '@solana/wallet-adapter-base';
+import { ConnectionProvider, WalletProvider } from '@solana/wallet-adapter-react';
+import { WalletDialogProvider, WalletMultiButton } from '@solana/wallet-adapter-material-ui';
+import {
+    LedgerWalletAdapter,
+    PhantomWalletAdapter,
+    SlopeWalletAdapter,
+    SolflareWalletAdapter,
+    SolletExtensionWalletAdapter,
+    SolletWalletAdapter,
+    TorusWalletAdapter,
+} from '@solana/wallet-adapter-wallets';
+import { clusterApiUrl } from '@solana/web3.js';
+import { useSnackbar } from 'notistack';
+import React, { FC, ReactNode, useCallback, useMemo } from 'react';
+import { Theme } from './Theme';
 
-// Use require instead of import, and order matters
-require('./App.css');
+// Default styles that can be overridden by your app
+require('@solana/wallet-adapter-react-ui/styles.css');
 
-const theme = createTheme({
-    palette: {
-        mode: 'dark',
-        primary: {
-            main: deepPurple[700],
-        },
-    },
-    components: {
-        MuiButtonBase: {
-            styleOverrides: {
-                root: {
-                    justifyContent: 'flex-start',
-                },
-            },
-        },
-        MuiButton: {
-            styleOverrides: {
-                root: {
-                    textTransform: 'none',
-                    padding: '12px 16px',
-                },
-                startIcon: {
-                    marginRight: 8,
-                },
-                endIcon: {
-                    marginLeft: 8,
-                },
-            },
-        },
-    },
-});
-
-const App: FC = () => {
+export const App: FC = () => {
     return (
-        <StyledEngineProvider injectFirst>
-            <ThemeProvider theme={theme}>
-                <SnackbarProvider>
-                    <Wallet />
-                </SnackbarProvider>
-            </ThemeProvider>
-        </StyledEngineProvider>
+        <Theme>
+            <Context>
+                <Content />
+            </Context>
+        </Theme>
     );
 };
 
-export default App;
+const Context: FC<{ children: ReactNode }> = ({ children }) => {
+    // The network can be set to 'devnet', 'testnet', or 'mainnet-beta'.
+    const network = WalletAdapterNetwork.Devnet;
+
+    // You can also provide a custom RPC endpoint.
+    const endpoint = useMemo(() => clusterApiUrl(network), [network]);
+
+    // @solana/wallet-adapter-wallets includes all the adapters but supports tree shaking and lazy loading --
+    // Only the wallets you configure here will be compiled into your application, and only the dependencies
+    // of wallets that your users connect to will be loaded.
+    const wallets = useMemo(
+        () => [
+            new PhantomWalletAdapter(),
+            new SlopeWalletAdapter(),
+            new SolflareWalletAdapter(),
+            new TorusWalletAdapter(),
+            new LedgerWalletAdapter(),
+            new SolletWalletAdapter({ network }),
+            new SolletExtensionWalletAdapter({ network }),
+        ],
+        [network]
+    );
+
+    const { enqueueSnackbar } = useSnackbar();
+    const onError = useCallback(
+        (error: WalletError) => {
+            enqueueSnackbar(error.message ? `${error.name}: ${error.message}` : error.name, { variant: 'error' });
+            console.error(error);
+        },
+        [enqueueSnackbar]
+    );
+
+    return (
+        <ConnectionProvider endpoint={endpoint}>
+            <WalletProvider wallets={wallets} onError={onError} autoConnect>
+                <WalletDialogProvider>{children}</WalletDialogProvider>
+            </WalletProvider>
+        </ConnectionProvider>
+    );
+};
+
+const Content: FC = () => {
+    return <WalletMultiButton />;
+};
