+++ packages/starter/material-ui-starter/src/index.tsx
@@ -1,13 +1,10 @@
 import React, { StrictMode } from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
-
-// Use require instead of import, and order matters
-require('./index.css');
+import { App } from './App';
 
 ReactDOM.render(
     <StrictMode>
         <App />
     </StrictMode>,
-    document.getElementById('root')
+    document.getElementById('app')
 );
